monitor_app/requirements.txt
redis
pymodbus
pyserial
psycopg2
paho-mqtt